// api/src/utils/interfaces/common.ts
export interface CreateFeedbackOnSlideDto {
  slideId: string;
  message: Express.Multer.File | string | null;
  isPublished?: boolean;
}
export interface TFeedbackOnSlideResponse {
  id: string;
  userId: string;
  slideId: string;
  message: Express.Multer.File | string | null;
  isPublished: boolean;
  createdAt: Date;
  updatedAt: Date;
}